requirements.txt
fastapi
uvicorn[standard]
sqlalchemy[asyncio]
aiosqlite
httpx[socks]
pyyaml
python-jose[cryptography]
passlib[bcrypt]
apscheduler
